+++ src/main/java/tech/araopj/jokes/request/entity/JokeRequestBody.java
@@ -1,4 +1,4 @@
-package tech.araopj.jokes.requests.entity;
+package tech.araopj.jokes.request.entity;
 
 import tech.araopj.jokes.entity.Category;
 import tech.araopj.jokes.entity.Language;
